backwards compatibility of gate-analyzer

+++ core/Solver.h
@@ -142,7 +142,7 @@ public:
   }
 
   void printProblem() {
-    for (int i = 0; i < clauses.size(); i++) {
+    for (size_t i = 0; i < clauses.size(); i++) {
       printClause2(clauses[i]);
       printf("\n");
     }
